packages/policy/pyproject.toml
[project]
name = "ai-platform-policy"
version = "0.1.0"
description = "Policy Engine - tool governance, PII enforcement, access control"
requires-python = ">=3.11"
dependencies = [
    "ai-platform-core",
    "pyyaml>=6.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_platform.policy*"]
